pyproject.toml
[tool.poetry]
name = "collator"
version = "0.1.0"
description = "Collator allows you to run your project with different hyperparameters and organize your results"
authors = ["Rory Bedford <rorybedford@protonmail.com>"]
license = "GPL3"
readme = "README.md"
packages = [{ include = "collator", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
collator = "collator.main:main"
